chore: move read write splitting files into a read_write_splitting directory (#609)
