+++ common/lib/plugins/read_write_splitting/read_write_splitting_plugin.ts
@@ -1,33 +1,37 @@
 /*
-  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- 
-  Licensed under the Apache License, Version 2.0 (the "License").
-  You may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
- 
-  http://www.apache.org/licenses/LICENSE-2.0
- 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import { AbstractConnectionPlugin } from "../abstract_connection_plugin";
-import { HostInfo, FailoverError, HostRole } from "../";
-import { PluginService } from "../plugin_service";
-import { HostListProviderService } from "../host_list_provider_service";
-import { OldConnectionSuggestionAction } from "../old_connection_suggestion_action";
-import { HostChangeOptions } from "../host_change_options";
-import { WrapperProperties } from "../wrapper_property";
-import { Messages } from "../utils/messages";
-import { logger } from "../../logutils";
-import { SqlMethodUtils } from "../utils/sql_method_utils";
-import { ClientWrapper } from "../client_wrapper";
-import { getWriter, logAndThrowError } from "../utils/utils";
-import { CanReleaseResources } from "../can_release_resources";
-import { PoolClientWrapper } from "../pool_client_wrapper";
+import { AbstractConnectionPlugin } from "../../abstract_connection_plugin";
+import {
+  HostInfo,
+  FailoverError,
+  HostRole
+} from "../../index";
+import { PluginService } from "../../plugin_service";
+import { HostListProviderService } from "../../host_list_provider_service";
+import { OldConnectionSuggestionAction } from "../../old_connection_suggestion_action";
+import { HostChangeOptions } from "../../host_change_options";
+import { WrapperProperties } from "../../wrapper_property";
+import { Messages } from "../../utils/messages";
+import { logger } from "../../../logutils";
+import { SqlMethodUtils } from "../../utils/sql_method_utils";
+import { ClientWrapper } from "../../client_wrapper";
+import { getWriter, logAndThrowError } from "../../utils/utils";
+import { CanReleaseResources } from "../../can_release_resources";
+import { PoolClientWrapper } from "../../pool_client_wrapper";
 
 export class ReadWriteSplittingPlugin extends AbstractConnectionPlugin implements CanReleaseResources {
   private static readonly subscribedMethods: Set<string> = new Set(["initHostProvider", "connect", "notifyConnectionChanged", "query"]);
